alteração do package repository para model, ccriacao da camada service e userService, adicionada dependencia openapi stater para usar o swagger.
