+++ src/main/java/com/projeto/redesocial/model/Post.java
@@ -1,4 +1,4 @@
-package com.projeto.redesocial.entity;
+package com.projeto.redesocial.model;
 
 import java.util.Date;
 
@@ -22,13 +22,16 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "likes", schema = "redesocial")
-public class Like {
+@Table(name = "posts", schema = "redesocial")
+public class Post {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "likes_sq")
-    @SequenceGenerator(schema = "redesocial", name = "likes_sq", sequenceName = "likes_sq", initialValue = 1, allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "posts_sq")
+    @SequenceGenerator(schema = "redesocial", name = "posts_sq", sequenceName = "posts_sq", initialValue = 1, allocationSize = 1)
     private Long id;
+
+    @Column(name = "text_content")
+    private String textContent;
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -39,11 +42,6 @@ public class Like {
     private Date updatedAt;
 
     @ManyToOne
-    @JoinColumn(name = "posts_id")
-    private Post post;
-
-    @ManyToOne
     @JoinColumn(name = "users_id")
     private User user;
-
 }
