+++ src/main/java/com/projeto/redesocial/model/PostPhoto.java
@@ -1,4 +1,4 @@
-package com.projeto.redesocial.entity;
+package com.projeto.redesocial.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
