+++ src/main/java/com/projeto/redesocial/repository/CommentRepository.java
@@ -1,7 +1,8 @@
 package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.projeto.redesocial.entity.Comment;
+
+import com.projeto.redesocial.model.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
