+++ src/main/java/com/projeto/redesocial/repository/FollowerRepository.java
@@ -2,7 +2,7 @@ package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.projeto.redesocial.entity.Follower;
+import com.projeto.redesocial.model.Follower;
 
 public interface FollowerRepository extends JpaRepository<Follower, Long> {
