+++ src/main/java/com/projeto/redesocial/repository/LikeRepository.java
@@ -2,7 +2,7 @@ package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.projeto.redesocial.entity.Like;
+import com.projeto.redesocial.model.Like;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
