+++ src/main/java/com/projeto/redesocial/repository/PostPhotoRepository.java
@@ -2,7 +2,7 @@ package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.projeto.redesocial.entity.PostPhoto;
+import com.projeto.redesocial.model.PostPhoto;
 
 public interface PostPhotoRepository extends JpaRepository<PostPhoto, Long> {
