+++ src/main/java/com/projeto/redesocial/repository/PostRepository.java
@@ -2,7 +2,7 @@ package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.projeto.redesocial.entity.Post;
+import com.projeto.redesocial.model.Post;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
