+++ src/main/java/com/projeto/redesocial/repository/TagRepository.java
@@ -2,7 +2,7 @@ package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.projeto.redesocial.entity.Tag;
+import com.projeto.redesocial.model.Tag;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
