+++ src/main/java/com/projeto/redesocial/repository/UserRepository.java
@@ -2,7 +2,7 @@ package com.projeto.redesocial.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.projeto.redesocial.entity.User;
+import com.projeto.redesocial.model.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
