+++ src/main/java/com/projeto/redesocial/service/UserService.java
@@ -1,0 +1,18 @@
+package com.projeto.redesocial.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.projeto.redesocial.model.User;
+import com.projeto.redesocial.repository.UserRepository;
+
+@Service
+public class UserService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    public User save(User user) {
+        return userRepository.save(user);
+    }
+}
